+++ src/Resources/app/administration/src/module/extension/sw-settings-payment/sw-settings-payment-detail/index.js
@@ -41,9 +41,9 @@ Component.override('sw-settings-payment-detail', {
 
         fetchMerchantIntegrations() {
             this.SwagPayPalApiCredentialsService
-                .getMerchantIntegrations()
+                .getMerchantInformation()
                 .then((response) => {
-                    this.merchantIntegrations = response;
+                    this.merchantIntegrations = response.merchantIntegrations ?? [];
                 });
         },
     },