+++ tests/Setting/Service/SettingsValidatorServiceTest.php
@@ -33,7 +33,7 @@ class SettingsValidatorServiceTest extends TestCase
         $validationService = new SettingsValidationService($systemSettings, new NullLogger());
         $validationService->validate();
         $validationService->validate(TestDefaults::SALES_CHANNEL);
-        static::assertTrue(true);
+        static::expectNotToPerformAssertions();
     }
 
     public function testValidateWithValidLiveDistinctSettings(): void
@@ -60,7 +60,7 @@ class SettingsValidatorServiceTest extends TestCase
         $validationService = new SettingsValidationService($systemSettings, new NullLogger());
         $validationService->validate();
         $validationService->validate(TestDefaults::SALES_CHANNEL);
-        static::assertTrue(true);
+        static::expectNotToPerformAssertions();
     }
 
     public function testValidateWithValidSandboxSettingsButSandboxDisabled(): void