+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 7.2.0
+- PPI-769 - Add Pay Later button to Express checkout shortcut
+- PPI-773 - The settings page is now structured in tabs
+
 # 7.1.0
 - PPI-679 - Pay upon invoice payment details are now shown in the order details
 - PPI-762 - Several payment methods are now available in non-PPCP markets